pyproject.toml
[project]
name = "gee-biophys"
version = "1.2.0"
description = "Custom time-series export of LAIe/FAPAR/FCOVER with YAML params."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Felix Specker", email = "felix.specker@wsl.ch" }
]
license = { text = "MIT" } 
dependencies = [
    "earthengine-api>=1.6.15",
    "geojson>=3.2.0",
    "loguru>=0.7.3",
    "numpy",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "scikit-learn==1.7.2",
    "typer>=0.12.0",
]

[project.scripts]
gee-biophys = "gee_biophys.cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# package your three code trees; skip non-code dirs
include = ["gee_biophys*", "models*", "train_pipeline*"]
exclude = ["notebooks*", "example_configs*", "test_configs*", "deprecated*", "geometries*", "auth*"]

[tool.setuptools.package-data]
"gee_biophys.models" = [
  "sl2p_params/*.json", 
  "specker_params/**/*.pkl",
  "specker_params/**/*.json",
]

[dependency-groups]
dev = [
    "pandas>=2.3.3",
    "pytest>=9.0.0",
    "ruff>=0.14.4",
]
# (optional) ensure data is included
[tool.setuptools]
include-package-data = true

